+++ py/CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ## Unreleased
 
+- Add `passive` flag to cardinality limits in project configs. ([#3199](https://github.com/getsentry/relay/pull/3199))
 - Don't run validation on renormalization, and don't normalize spans. ([#3214](https://github.com/getsentry/relay/pull/3214))
 
 ## 0.8.46